ClimateOBS script updates

+++ scripts/python/DataMart_hourly_download.py
@@ -16,40 +16,45 @@ def isnumber(x):
     except:
         return False
 
-def objstore_to_df(objpath):
+def objstore_to_df(objpath,onprem=False):
     filename = objpath.split("/")[-1]
     filetype = filename.split(".")[-1]
-    local_folder = 'raw_data/temp_file'
-    if not os.path.exists(local_folder):
-        os.makedirs(local_folder)
-
-    local_path = os.path.join(local_folder,filename)
-    ostore.get_object(local_path=local_path, file_path=objpath)
+    if onprem:
+        local_path = objpath
+    else:
+        local_folder = 'raw_data/temp_file'
+        if not os.path.exists(local_folder):
+            os.makedirs(local_folder)
+        local_path = os.path.join(local_folder,filename)
+        ostore.get_object(local_path=local_path, file_path=objpath)
     match filetype:
         case 'csv':
             output = pd.read_csv(local_path)
         case 'parquet':
             output = pd.read_parquet(local_path)
-    os.remove(local_path)
+    if not onprem:
+        os.remove(local_path)
 
     return output
 
-def df_to_objstore(df, objpath):
+def df_to_objstore(df, objpath, onprem=False):
     filename = objpath.split("/")[-1]
     filetype = filename.split(".")[-1]
-    local_folder = 'raw_data/temp_file'
-    if not os.path.exists(local_folder):
-        os.makedirs(local_folder)
-
-    local_path = os.path.join(local_folder,filename)
+    if onprem:
+        local_path = objpath
+    else:
+        local_folder = 'raw_data/temp_file'
+        if not os.path.exists(local_folder):
+            os.makedirs(local_folder)
+        local_path = os.path.join(local_folder,filename)
     match filetype:
         case 'csv':
             df.to_csv(local_path)
         case 'parquet':
             df.to_parquet(local_path)
-    
-    ostore.put_object(local_path=local_path, ostore_path=objpath)
-    os.remove(local_path)
+    if not onprem:
+        ostore.put_object(local_path=local_path, ostore_path=objpath)
+        os.remove(local_path)
 
 
 #Search for desired variables in xml file, grab associated values:
@@ -64,8 +69,9 @@ def retrieve_xml_values(xml_file,var_names):
     return output
 
 class data_config():
-    def __init__(self,objfolder, stn_list, src_stn_list, url_template, fname_template, var_names):
+    def __init__(self, objfolder, onprem, stn_list, src_stn_list, url_template, fname_template, var_names):
         self.objfolder = objfolder
+        self.onprem = onprem
         self.stn_list = stn_list
         self.src_stn_list = src_stn_list
         self.url_template = url_template
@@ -79,16 +85,19 @@ class data_config():
         #Conver dt_range to UTC since data on datamart is in UTC:
         #Limit dt_range_utc to < current time so that it is not trying to grab non-existant data:
         dt_range_utc = dt_range[0:date.hour+1] + datetime.timedelta(hours=8)
-        
+
         all_data_objpath = os.path.join(self.objfolder,f'{dt_txt}.parquet')
-        all_data_objs = ostore.list_objects(self.objfolder,return_file_names_only=True)
+        if self.onprem == False:
+            all_data_objs = ostore.list_objects(self.objfolder,return_file_names_only=True)
+        else:
+            all_data_objs = os.listdir(self.objfolder)
         local_file_path = 'raw_data/temp_data'
         if not os.path.exists(local_file_path):
             os.makedirs(local_file_path)
         if all_data_objpath in all_data_objs:
             #ostore.get_object(local_path=local_data_fpath, file_path=all_data_objpath)
             #output = pd.read_parquet(local_data_fpath)
-            output = objstore_to_df(all_data_objpath)
+            output = objstore_to_df(all_data_objpath,self.onprem)
         else:
             output_ind = pd.MultiIndex.from_product([self.src_stn_list,dt_range], names=["Station", "DateTime"])
             output = pd.DataFrame(data=None,index=output_ind,columns=self.var_names+['f_read'])
@@ -100,7 +109,7 @@ class data_config():
                 #Format html string for data location:
                 dt_str = dt.strftime('%Y-%m-%d-%H00')
                 date_str = dt.strftime(default_date_format)
-                
+
                 remote_location = self.url_template.format(date_str=date_str)
                 if type(self.fname_template)!=list:
                     self.fname_template=[self.fname_template]
@@ -119,7 +128,7 @@ class data_config():
                             #If file location exists, proceed with file download:
                             if r.status_code == requests.codes.ok:
                                 with open(local_filename, 'wb') as f:
-                                    for chunk in r.iter_content(chunk_size=8192): 
+                                    for chunk in r.iter_content(chunk_size=8192):
                                         f.write(chunk)
                                 break
                     #If file was downloaded succefully, write variables to dataframe:
@@ -128,17 +137,17 @@ class data_config():
                         #Set f_read to True so script does not re-download file in subsequent runs:
                         output.loc[stn,dt - datetime.timedelta(hours=8)].iloc[-1] = True
                         os.remove(local_filename)
-                
+
         #Save dataframe to parquet file and send to object store:
         #output.to_parquet(local_data_fpath)
         #ostore.put_object(local_path=local_data_fpath, ostore_path=all_data_objpath)
-        df_to_objstore(output, all_data_objpath)
-    
+        df_to_objstore(output, all_data_objpath, self.onprem)
+
     def write_data(self,date,objpath,src_varname,out_varname):
         dt_txt = date.strftime('%Y%m%d')
         all_data_objpath = os.path.join(self.objfolder,f'{dt_txt}.parquet')
         #Reformat data for each variable into their own dataframes:
-        data = objstore_to_df(all_data_objpath)
+        data = objstore_to_df(all_data_objpath,self.onprem)
         output = data.loc[:,src_varname].unstack(0)[self.src_stn_list]
 
         #Remove non-numeric (and nan) values:
@@ -151,7 +160,7 @@ class data_config():
         output.columns = self.stn_list
 
         #Save temperature/precip dataframes as csv files and send to object store:
-        df_to_objstore(output,output_obj)
+        df_to_objstore(output,output_obj,self.onprem)
 
 
 if __name__ == '__main__':
@@ -165,7 +174,7 @@ if __name__ == '__main__':
     #If downloading past days, set hour to 23 so entire day is downloaded (scripts only attempts download up to current hour for present day):
     if days_back>0:
         current_date.replace(hour=23)
-    
+
     #Grab list of stations to download data for from object store:
     #This is the updated station list for ClimateOBS
     stn_list_local = 'raw_data/ECCC_stationlist.csv'
@@ -179,12 +188,12 @@ if __name__ == '__main__':
     url_template = 'http://hpfx.collab.science.gc.ca/{date_str}/WXO-DD/observations/swob-ml/{date_str}/'
     #Template for file names:
     fname_template = ['{stn}/{dt_str}-{stn}-MAN-swob.xml','{stn}/{dt_str}-{stn}-AUTO-swob.xml']
-    
+
     #List of variables to grab data for:
     var_names = ['air_temp','avg_air_temp_pst1hr','pcpn_amt_pst1hr']
     objfolder = 'RFC_DATA/ECCC/hourly/parquet/'
 
-    ECCC = data_config(objfolder,stn_list,src_stn_list,url_template,fname_template,var_names)
+    ECCC = data_config(objfolder,False,stn_list,src_stn_list,url_template,fname_template,var_names)
     ECCC.update_data(current_date)
 
     objpath = 'RFC_DATA/ECCC/hourly/csv/'
@@ -196,7 +205,7 @@ if __name__ == '__main__':
     crd_url_template = 'http://hpfx.collab.science.gc.ca/{date_str}/WXO-DD/observations/swob-ml/partners/bc-crd/{date_str}/'
     crd_fname_template = ['{stn}/{dt_str}-bc-crd-{stn}-{stn}-AUTO-swob.xml']
     crd_var_names = ['air_temp','pcpn_amt_pst1hr']
-    CRD = data_config(crd_objfolder,crd_stn_list,crd_stn_list,crd_url_template,crd_fname_template,crd_var_names)
+    CRD = data_config(crd_objfolder,False,crd_stn_list,crd_stn_list,crd_url_template,crd_fname_template,crd_var_names)
     CRD.update_data(current_date)
 
     objpath = 'RFC_DATA/CRD/csv/'
